Add option for H5 dataset compression
Added a CLI option to enable compression of the H5 datasets. Currently
applies compression to all datasets and supports only two options
built-in to h5py: gzip and lzf

So far no noticeable performance hit when running ndlar-flow, but
seeing 75% (lzf) to 85% (gzip) compression ratios.

May eventually implement the compression option in the YAML config, but
this works well for now.